+++ plop-templates/components/js.hbs
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import {{lowerCase name}} from './{{lowerCase name}}.vue';
+import {{lowerCase name}} from './press-{{lowerCase name}}.vue';
 import PressComponentBase from '../../press-component';
 
 Vue.component('{{lowerCase name}}', {{lowerCase name}});
