+++ plopfile.js
@@ -12,12 +12,13 @@ module.exports = function(plop) {
     actions: [
       {
         type: 'add',
-        path: 'src/components/{{kebabCase name}}/index.js',
+        path: 'src/components/press-{{kebabCase name}}/index.js',
         templateFile: 'plop-templates/components/js.hbs'
       },
       {
         type: 'add',
-        path: 'src/components/{{kebabCase name}}/{{kebabCase name}}.vue',
+        path:
+          'src/components/press-{{kebabCase name}}/press-{{kebabCase name}}.vue',
         templateFile: 'plop-templates/components/vue.hbs'
       },
       {
@@ -32,7 +33,7 @@ module.exports = function(plop) {
         path: 'src/index.js',
         pattern: /(\/\/ PLOP: END COMPONENT IMPORT)/gi,
         template:
-          "import {{pascalCase name}} from './components/{{kebabCase name}};\n$1"
+          "import {{pascalCase name}} from './components/press-{{kebabCase name}};\n$1"
       }
     ]
   });
